[TS-330] Refactor: my info page button change to common menu button
